+++ src/Day2/Practice/SebahattinTask3.java
@@ -1,0 +1,43 @@
+package Day2.Practice;
+
+public class SebahattinTask3 {
+
+        public static void main(String[] args) {
+            char S = 83;
+            char e = 101;
+            char b = 98;
+            char a = 97;
+            char h = 104;
+            char t = 116;
+            char i = 105;
+            char n = 110;
+            char space = 32;
+            char B = 66;
+            char E = 69;
+            char Y = 89;
+            char C = 67;
+            char U = 85;
+            char R = 82;
+
+            System.out.print(S);
+            System.out.print(e);
+            System.out.print(b);
+            System.out.print(a);
+            System.out.print(h);
+            System.out.print(a);
+            System.out.print(t);
+            System.out.print(t);
+            System.out.print(i);
+            System.out.print(n);
+            System.out.print(space);
+            System.out.print(B);
+            System.out.print(E);
+            System.out.print(Y);
+            System.out.print(C);
+            System.out.print(U);
+            System.out.print(R);
+
+        }
+    }
+
+
